refactor: extract UI component
