+++ ui/ui.go
@@ -2,182 +2,99 @@ package ui
 
 import (
 	"fmt"
-	"strings"
+	"os"
+	"time"
 
-	human "github.com/dustin/go-humanize"
-	net "github.com/libp2p/go-libp2p-core/network"
-	ma "github.com/multiformats/go-multiaddr"
+	"github.com/dustin/go-humanize"
+	"github.com/libp2p/hydra-booster/node"
+	"github.com/libp2p/hydra-booster/reports"
 )
 
-// ...
 const (
-	QClrLine = "\033[K"
-	QReset   = "\033[2J"
+	logey = iota
+	gooey
 )
 
-/*
-Move the cursor up N lines:
-  \033[<N>A
-- Move the cursor down N lines:
-  \033[<N>B
-- Move the cursor forward N columns:
-  \033[<N>C
-- Move the cursor backward N columns:
-  \033[<N>D
-*/
+func NewUI(nodes []*node.HydraNode, statusReports chan reports.StatusReport, start time.Time) error {
+	uiType := logey
 
-// ...
-const (
-	Clear     = 0
-	LightBlue = 94
-)
-
-// ...
-const (
-	Black = 30 + iota
-	Red
-	Green
-	Yellow
-	Blue
-	Magenta
-	Cyan
-	LightGray
-)
-
-func color(color int, s string) string {
-	return fmt.Sprintf("\x1b[%dm%s\x1b[0m", color, s)
-}
-
-const width = 25
-
-func padPrint(line int, label, value string) {
-	putMessage(line, fmt.Sprintf("%s%s%s", label, strings.Repeat(" ", width-len(label)), value))
-}
-
-func putMessage(line int, mes string) {
-	fmt.Printf("\033[%d;0H%s%s", line, QClrLine, mes)
-}
-
-func printDataSharedLine(line int, bup int, totup int64, rateup float64) {
-	pad := "            "
-	a := fmt.Sprintf("%d            ", bup)[:12]
-	b := (human.Bytes(uint64(totup)) + pad)[:12]
-	c := (human.Bytes(uint64(rateup)) + "/s" + pad)[:12]
-
-	padPrint(line, "", a+b+c)
-}
-
-// GooeyApp ..
-type GooeyApp struct {
-	Title      string
-	DataFields []*DataLine
-	Log        *Log
-}
-
-// Print ...
-func (a *GooeyApp) Print() {
-	fmt.Println(QReset)
-	putMessage(1, a.Title)
-	for _, dl := range a.DataFields {
-		dl.Print()
-	}
-	a.Log.Print()
-}
-
-// NewDataLine ...
-func (a *GooeyApp) NewDataLine(line int, label, defval string) *DataLine {
-	dl := &DataLine{
-		Default: defval,
-		Label:   label,
-		Line:    line,
-	}
-	a.DataFields = append(a.DataFields, dl)
-
-	return dl
-}
-
-// DataLine ...
-type DataLine struct {
-	Label   string
-	Line    int
-	Default string
-	LastVal string
-}
-
-// SetVal ...
-func (dl *DataLine) SetVal(s string) {
-	dl.LastVal = s
-	dl.Print()
-}
-
-// Print ...
-func (dl *DataLine) Print() {
-	s := dl.Default
-	if dl.LastVal != "" {
-		s = dl.LastVal
+	if len(nodes) == 0 {
+		return fmt.Errorf("ui needs at least one node")
 	}
 
-	padPrint(dl.Line, dl.Label, s)
-}
-
-// Log ...
-type Log struct {
-	Size      int
-	StartLine int
-	Messages  []string
-	beg       int
-	end       int
-}
-
-// NewLog ...
-func NewLog(line, size int) *Log {
-	return &Log{
-		Size:      size,
-		StartLine: line,
-		Messages:  make([]string, size),
-		end:       -1,
+	if len(nodes) == 1 {
+		uiType = gooey
 	}
-}
 
-// Add ...
-func (l *Log) Add(m string) {
-	l.end = (l.end + 1) % l.Size
-	if l.Messages[l.end] != "" {
-		l.beg++
+	switch uiType {
+	case logey: // many node
+		for {
+			r, ok := <-statusReports
+			if !ok {
+				break
+			}
+			printStatusLine(r, start)
+		}
+	case gooey: // 1 node
+		ga := &GooeyApp{Title: "Hydra Booster Node", Log: NewLog(15, 15)}
+		ga.NewDataLine(3, "Peer ID", nodes[0].Host.ID().Pretty())
+		econs := ga.NewDataLine(4, "Connections", "0")
+		uniqprs := ga.NewDataLine(5, "Unique Peers Seen", "0")
+		emem := ga.NewDataLine(6, "Memory Allocated", "0MB")
+		eprov := ga.NewDataLine(7, "Stored Provider Records", "0")
+		eprlat := ga.NewDataLine(8, "Store Provider Latency", "0s")
+		etime := ga.NewDataLine(9, "Uptime", "0h 0m 0s")
+		ga.Print()
+
+		second := time.NewTicker(time.Second)
+
+		for {
+			select {
+			// case m := <-messages:
+			// 	ga.Log.Add(m)
+			// 	ga.Log.Print()
+			case r, ok := <-statusReports:
+				if !ok {
+					second.Stop()
+					statusReports = nil
+					break
+				}
+				emem.SetVal(humanize.Bytes(r.MemStats.Alloc))
+				econs.SetVal(fmt.Sprintf("%d peers", r.TotalConnectedPeers))
+				uniqprs.SetVal(fmt.Sprint(r.TotalUniquePeers))
+				eprov.SetVal(fmt.Sprint(r.TotalProvs))
+
+				if r.TotalProvs > 0 {
+					eprlat.SetVal(fmt.Sprint(r.TotalProvTime / time.Duration(r.TotalProvs)))
+				}
+
+				ga.Print()
+			case <-second.C:
+				t := time.Since(start)
+				h := int(t.Hours())
+				m := int(t.Minutes()) % 60
+				s := int(t.Seconds()) % 60
+				etime.SetVal(fmt.Sprintf("%dh %dm %ds", h, m, s))
+				ga.Print()
+			}
+		}
+	default:
+		return fmt.Errorf("unknown UI type %v", uiType)
 	}
-	l.Messages[l.end] = m
+
+	return nil
 }
 
-// Print ...
-func (l *Log) Print() {
-	for i := 0; i < l.Size; i++ {
-		putMessage(l.StartLine+i, l.Messages[(l.beg+i)%l.Size])
-	}
+func printStatusLine(report reports.StatusReport, start time.Time) {
+	fmt.Fprintf(
+		os.Stderr,
+		"[NumDhts: %d, Uptime: %s, Memory Usage: %s, TotalPeers: %d/%d, Total Provs: %d, BootstrapsDone: %d]\n",
+		report.TotalHydraNodes,
+		time.Second*time.Duration(int(time.Since(start).Seconds())),
+		humanize.Bytes(report.MemStats.Alloc),
+		report.TotalConnectedPeers,
+		report.TotalUniquePeers,
+		report.TotalProvs,
+		report.TotalBootstrappedHydraNodes,
+	)
 }
-
-// LogNotifee ...
-type LogNotifee struct {
-	addMes chan<- string
-}
-
-// Listen ...
-func (ln *LogNotifee) Listen(net.Network, ma.Multiaddr) {}
-
-// ListenClose ...
-func (ln *LogNotifee) ListenClose(net.Network, ma.Multiaddr) {}
-
-// Connected ...
-func (ln *LogNotifee) Connected(_ net.Network, c net.Conn) {
-	ln.addMes <- fmt.Sprintf("New connection from %s", c.RemotePeer().Pretty())
-}
-
-// Disconnected ...
-func (ln *LogNotifee) Disconnected(_ net.Network, c net.Conn) {
-	ln.addMes <- fmt.Sprintf("Lost connection to %s", c.RemotePeer().Pretty())
-}
-
-// OpenedStream ...
-func (ln *LogNotifee) OpenedStream(net.Network, net.Stream) {}
-
-// ClosedStream ...
-func (ln *LogNotifee) ClosedStream(net.Network, net.Stream) {}
